refactor(tui): move render core + model into agent_tui_kit (Phase 3)
Complete the mechanical extraction of the TUI kit (Phase 3 of the
component-library plan). All pure rendering and state now live in
agent_tui_kit (133 tests); tui re-exports them.

- render core: render_md, pulldown, mermaid_sequence, util, renderable,
  log_column
- cells: markdown (MarkdownCell), separator, text, thinking, tool
- state: log (LogItem/LogItemKind/SystemMsgStyle + LogCoordinator),
  thinking (blocks/state/popup), selection (PopupTextSelection)
- widgets: tool_widget (ToolWidget/ToolPhase/ToolRenderOutput)
- cells/markdown App-integration tests split into tui (markdown_integration_tests)
- Ctx design doc + log_column move (Ctx step 1)

Tests: tui 471 + kit 133 = 604 (baseline), tact-ui 45. No behavior change.
