+++ crates/agent_tui_kit/src/state/thinking.rs
@@ -3,22 +3,22 @@ use std::{
     time::{Duration, Instant},
 };
 
-use super::tool_state::PopupTextSelection;
+use super::selection::PopupTextSelection;
 
 /// Streaming thinking content anchored at one shared-log placeholder row.
 #[derive(Debug, Clone)]
-pub(crate) struct ActiveThinkingBlock {
-    pub(crate) phys_idx: usize,
-    pub(crate) content: String,
+pub struct ActiveThinkingBlock {
+    pub phys_idx: usize,
+    pub content: String,
     pending_line: String,
     completed_tail: VecDeque<String>,
-    pub(crate) started_at: Instant,
+    pub started_at: Instant,
 }
 
 impl ActiveThinkingBlock {
     const MAX_TAIL_LINES: usize = 3;
 
-    pub(crate) fn new(phys_idx: usize, started_at: Instant) -> Self {
+    pub fn new(phys_idx: usize, started_at: Instant) -> Self {
         Self {
             phys_idx,
             content: String::new(),
@@ -28,7 +28,7 @@ impl ActiveThinkingBlock {
         }
     }
 
-    pub(crate) fn push_delta(&mut self, delta: &str) {
+    pub fn push_delta(&mut self, delta: &str) {
         self.content.push_str(delta);
         self.pending_line.push_str(delta);
 
@@ -42,7 +42,7 @@ impl ActiveThinkingBlock {
         }
     }
 
-    pub(crate) fn display_tail(&self) -> Vec<String> {
+    pub fn display_tail(&self) -> Vec<String> {
         let mut tail: Vec<_> = self.completed_tail.iter().cloned().collect();
         if !self.pending_line.is_empty() {
             tail.push(self.pending_line.clone());
@@ -53,41 +53,41 @@ impl ActiveThinkingBlock {
         tail
     }
 
-    pub(crate) fn body_line_count(&self) -> usize {
+    pub fn body_line_count(&self) -> usize {
         self.display_tail().len().clamp(1, Self::MAX_TAIL_LINES)
     }
 
-    pub(crate) fn is_blank(&self) -> bool {
+    pub fn is_blank(&self) -> bool {
         self.content.trim().is_empty()
     }
 }
 
 /// Thinking state: one active direct card, completed cards, and the detail popup.
 #[derive(Default)]
-pub(crate) struct ThinkingState {
+pub struct ThinkingState {
     /// Reasoning card currently receiving streaming deltas.
-    pub(crate) active: Option<ActiveThinkingBlock>,
+    pub active: Option<ActiveThinkingBlock>,
     /// Completed reasoning cards, retained for rendering and detail popups.
-    pub(crate) blocks: Vec<ThinkingBlock>,
+    pub blocks: Vec<ThinkingBlock>,
     /// Detail popup state.
-    pub(crate) popup: Option<ThinkingPopup>,
+    pub popup: Option<ThinkingPopup>,
 }
 
 /// A completed reasoning card anchored at one shared-log placeholder row.
 #[derive(Debug, Clone)]
-pub(crate) struct ThinkingBlock {
-    pub(crate) phys_idx: usize,
-    pub(crate) content: String,
-    pub(crate) summary: String,
+pub struct ThinkingBlock {
+    pub phys_idx: usize,
+    pub content: String,
+    pub summary: String,
     /// Cached Markdown rendered lines, used for popup display, avoiding per-frame re-rendering.
-    pub(crate) cached_markdown: Vec<ratatui::text::Line<'static>>,
+    pub cached_markdown: Vec<ratatui::text::Line<'static>>,
     /// Duration of the thinking phase.
-    pub(crate) elapsed: Duration,
+    pub elapsed: Duration,
 }
 
 /// Thinking popup state.
 #[derive(Debug, Clone)]
-pub(crate) struct ThinkingPopup {
+pub struct ThinkingPopup {
     /// Stable shared-log placeholder index for active or completed content.
     pub phys_idx: usize,
     pub title: String,
@@ -100,7 +100,7 @@ pub(crate) struct ThinkingPopup {
 }
 
 impl ThinkingPopup {
-    pub(crate) fn copy_content(&self, full_content: &str) -> String {
+    pub fn copy_content(&self, full_content: &str) -> String {
         self.selection
             .and_then(|selection| selection.normalized_non_empty(&self.selection_text))
             .map(|range| self.selection_text[range].to_string())
