+++ crates/agent_tui_kit/src/widgets/mod.rs
@@ -1,0 +1,5 @@
+//! Widget builders shared by the kit (tool card, popups, …).
+//!
+//! Phase 3 moves these out of `crates/tui/src/widgets` in cluster order.
+
+pub mod tool_widget;
