+++ crates/tui/src/render/cells/mod.rs
@@ -1,6 +1,16 @@
 pub(crate) mod code;
-pub(crate) mod markdown;
-pub(crate) mod separator;
-pub(crate) mod text;
-pub(crate) mod thinking;
-pub(crate) mod tool;
+pub(crate) mod separator {
+    pub(crate) use agent_tui_kit::render::cells::separator::*;
+}
+pub(crate) mod text {
+    pub(crate) use agent_tui_kit::render::cells::text::*;
+}
+pub(crate) mod thinking {
+    pub(crate) use agent_tui_kit::render::cells::thinking::*;
+}
+pub(crate) mod tool {
+    pub(crate) use agent_tui_kit::render::cells::tool::*;
+}
+
+#[cfg(test)]
+mod markdown_integration_tests;
