+++ crates/tui/src/widgets/mod.rs
@@ -2,4 +2,6 @@ pub(crate) mod help_widget;
 pub(crate) mod popup_widget;
 pub(crate) mod select_popup_widget;
 pub(crate) mod state;
-pub(crate) mod tool_widget;
+pub(crate) mod tool_widget {
+    pub(crate) use agent_tui_kit::widgets::tool_widget::*;
+}
